refactor(guide): split up steps into seperate files
