+++ lib/cmds/guide/helpers.js
@@ -1,0 +1,3 @@
+const guideOptimalColumns = 120
+export const GUIDE_MAX_WIDTH = process.stdout.columns < guideOptimalColumns ? process.stdout.columns : guideOptimalColumns
+export class AbortedError extends Error { }
